CompletableFutureProxy
